Revert "feat : register"
